Moving SE layer to ops.
